+++ build.sh
@@ -4,7 +4,7 @@
 
 set -e
 
-VERSION="0.4.0"
+VERSION="0.5.0"
 APP_NAME="kcsi"
 BUILD_DIR="bin"